[roll] Roll fuchsia [bazel_sdk] do not package libcxx when statically linking
When a user specifies that they want to statically link c++ via the
static_cpp_standard_library feature we want to avoid packaging the
libcxx dist libraries. This adds that check and removes the dist targets
in the fuchsia_wrap_cc_binary macro.

In the future we will need to move this logic into the dist target
itself.

Additionally, this CL adds new tests to verify that a package that
contains a mix of statically linked libraries and binaries that are not
statically linked still include the dist files.

Original-BUG: 368434718
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122625
Original-Revision: 2a3bfbee2338300e6724ff8385aa88f104b0ddf0
GitOrigin-RevId: 91e0850cc64831074484e53c552e07c4bd32ee1b
Change-Id: I812cbe956dbceb1fcb7759c928dfd05b61289397