[roll] Roll fuchsia [kernel][x86] lapic: Improve TSC-Deadline Timer support
On certain x86 platforms, the Local APIC timer can be used
in 'TSC Deadline' mode; in this mode, the timer is a one-shot
source that is armed by writing an absolute TSC deadline to
an MSR. When the TSC is equal to or beyond that deadline, the
timer is reset and an interrupt is raised. This mode is
particularly efficient when operating systems only want one-shot
timers; arming requires one MSR write.

Fuchsia enabled and used this mode, but there were
inefficiencies in how it did so; improve the efficiency and
use of this timer mode.
1) When arming the TSC deadline timer, Fuchsia always
   wrote to the Local APIC LVT_TIMER register to put the
   APIC into TSC Deadline mode and then used the TSC deadline
   timer; this was unnecessary, since the timer retains its
   mode until reconfigured.

2) When canceling a timer, which Fuchsia does often, the
   system would cancel both the TSC deadline timer (if any)
   and any conventional Local APIC timers. When in TSC
   deadline mode, there should not be any Local APIC timers
   armed, so skip disarming them.

This CL unconditionally enables TSC deadline mode in the lapic
when it is supported, as the lapic (arch) layer is initialized
before the timer (platform) layer, but if use_tsc_deadline is
false at the platform layer, enabling a one-shot timer via
the legacy path will disable TSC Deadline mode.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/632091
Original-Revision: d95baec97a7e412096a05d5398c75c0b97799aaa
GitOrigin-RevId: 154a5bc56818a753c9f92a6705b84e272467c252
Change-Id: I9cf5e1516861e7d02cceb10d5487956af1489aab