[roll] Roll fuchsia [starnix][touch-power-policy] Correct device metadata
The metadata for this device is incorrect and harmful. We cannot grab
minor device 0 for the MISC major device. The device number is assigned
to the Logitech bus mouse.

Instead, we should register this starnix-specific device as a dynamic
device, which assigns it a device number that has been reserved for
unknown devices.

Also, because the device is no longer in the MISC major device, this CL
switches the registration over to the "starnix" class, which is also
likely to be incorrect, but it's the class we've been using for these
Starnix-specific devices for the time being.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122992
Original-Revision: 2ed3143d71d6ae4aad85cec60ee0cdffa00817ee
GitOrigin-RevId: 6ca360df685c50bf92eb2a21d7a0ab4a80286f1f
Change-Id: Ie9a63146d7bf0a8e6dca3d601d3969f233160602