[roll] Roll fuchsia [cmdline] keep local console enabled with kernel.shell=true set
In the past, when setting kernel.shell=true the user shell was disabled
because the system continued to start user space and only one shell was
started (the kernel's). A while back it was changed such that enabling
the kernel shell stops the boot process before starting user space so no
need to disable the user shell.

Remove this particular interlock which now allows the user to drop into
the kernel shell, run some commands, and then exit the kernel shell and
continue a full user space with a user shell.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121458
Original-Revision: 07bd6bd7857b9d550c637614a83356c42d1663e8
GitOrigin-RevId: 5431210363d6886ff8d3af3734393c5f162ee4a1
Change-Id: I97130dd2ce4880bf4b1222877ebdb647dbe9f3e0