[roll] Roll fuchsia [kernel][target] remove the entire vestigial target/ directory
Forever ago this held the per target (soc/board/etc) layer of the
kernel, which was generally refactored into generic platforms when the
kernel was refactored to be per arch and not per board.

The dir was left around to hold the old boot shims, which have been
replaced with a new style boot shim.

Original-Bug: 42147821

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1121713
Original-Revision: 1bbe5b60ac22c264571e66e8118d46477d7917ac
GitOrigin-RevId: c2746002e99fd5e245246f63dfa504941471e158
Change-Id: Ie42847df87a16162cf3d31024cd0045c4cb1f5b7