[roll] Roll fuchsia [sdk][build] Enforce meta.stable in sdk_atom()
Require that `meta.stable` is specified for atom types that support
unstable atoms and that it is set to an appropriate value for the SDK
category. This requires adding `stable` to the `meta` in
`fidl_library()`. Ensure that `stable` is not passed to the static and
shared library templates as an additional precaution.

Similarly, ensure an `api` file is specified except when the atom is
unstable, including when it is in the SDK category "internal".

Also note that the value of `stable` in this build manifest does not
affect the IDK, though it should match the value in the IDK meta.json.

Original-Bug: 366221510, 364727245
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1122231
Original-Revision: 364fa025cc9a46950de72c1c5cef04a0c849c5d2
GitOrigin-RevId: 279d306f803db99400cfc249a897414e4c74fda8
Change-Id: I95468b8abb2dff048657c85650497eaac29a1fcf