[roll] Roll fuchsia [build] Move build_only_labels to top-level //:build_only group and use with subbuilds instead of universe
Universe is the wrong target to use for sub-builds, as it's
specifically a dependency of the product being compiled,
and sub-builds are really trying to do a build-only type of
operation.  We already have a build_only_labels mechanism,
although it's somewhat hidden.

This moves the build_only_labels mechanism into //BUILD.gn
and creates a top-level group for it to use that will play
nicely with the GN --root-pattern mechanism.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123882
Original-Revision: d0ac85199c0f8b1da133e9c1ebb786b5f0eefc8e
GitOrigin-RevId: b72db6976eaeb6d898c92f754c04a3c87d1a256d
Change-Id: I5aeea3de105309c7ceafc04e72e638f612dd1e96