[roll] Roll third_party/pigweed pw_allocator: Store TestHarness::Allocation inline
This module include a TestHarness class which can generate requests for
an allocator to allocate, reallocate and deallocate regions of memory.
Previously, this harness took a vector of Allocation structs that was
used to stored the details of these requests. This allowed, for example,
the test harness to deallocate memory that had previously been
allocated.

Requiring this vector could result in significant overhead if a
substantial number of allocations were to be tested at once. This
overhead is unecessary though, as there is another place to store the
request details: in the allocations themselves!

This change sets a minimum size and alignment on requests. It makes the
Allocation struct an IntrusiveList item, and stores the Allocations in
the allocated memory.

Original-Reviewed-on: https://pigweed-review.googlesource.com/c/pigweed/pigweed/+/230894
Presubmit-Verified: CQ Bot Account <pigweed-scoped@luci-project-accounts.iam.gserviceaccount.com>
Original-Revision: a4dab0be18a649c79759880308ed051b0a2576a4
GitOrigin-RevId: bc03729ec989e5b3ad9d3f9d4892b2bd96801eb5
Change-Id: I4fc3bc12399d489b98ab70c25c9eacbe0e3f4c78