[roll] Roll fuchsia [dl] Persist a RuntimeModule's module tree
This CL will store a list of the breadth-first order module tree
(inclusive of the root module) in the RuntimeModule for the dlopen-ed
module. This list is currently passed to relocations as is, but in the
future global modules will be prepended to this list for relocations.

This list is used directly by dlsym for symbol lookup.

This CL does not yet set the module tree for a module that was loaded
as a dependency of another module, which will be resolved in a future
CL.

Many tests that were skipped with kDlSymSupportsDeps are now enabled on
libdl. The kDepModuleHasDepTree feature flag is introduced to skip test
scenarios that exercise the unsupported case.

Original-Bug: 354786114
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116692
Original-Revision: 06f7df89df859b5e6cfe99b28701177bfbfe4255
GitOrigin-RevId: ace9258e04248b874423610ae3fc607c34b76ca3
Change-Id: Ie926bf7b936fa06b83b803d65d0c998b7ba953e7