[roll] Roll fuchsia [dl] Persist a RuntimeModule's direct deps
This CL will store a list of references to only the direct dependencies
of the RuntimeModule. This list is populated in the first BFS traversal
of the sessions_modules_ list, when direct dependencies are enqueued.
This list contains references to new dep modules that were loaded with
the root, as well as previously loaded dep modules.

EnqueueModule will now return a pointer to the loaded or new
RuntimeModule that was created while iterating over the session_modules_
list so that it can be added to the parent module's direct_deps list.

This list is used in the second BFS traversal of the root
module's module tree. Iterating over this list replaces the previous
`enqueue_deps` algorithm that checked whether a dep module was new to
the session or previously loaded in the second loop.

Original-Bug: 354786114
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1116532
Original-Revision: 518841d65b422d2eaacb9b94e781d324ce88c4cb
GitOrigin-RevId: 97a0c2be2c5a04484f8c4c8de3eff7afe6ec4a5e
Change-Id: Ifbe4223192ae71666522e2dd8b6cc6dac11b52cc