[roll] Roll fuchsia [vsock] Move bind state to server
The fidl protocol was split such that there is a channel connection per
socket. This simplifies some state management. Additionally, Listen2 was
split into 2 separate calls, Bind on the Connector and Listen on the
Listener. This split allows the vsock_service component to maintain
socket state entirely on the server. This is helpful for the case in
which we need to duplicate or migrate the socket to another process.
Neither of these operations are yet supported, but they can be more
easily supported after this change.

Original-Fixed: 368645901
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123881
Original-Revision: bd70c10a041d854be6841f623d48385d0370f261
GitOrigin-RevId: 6a9abb080de4c37994462288fdb10b2f767d6e72
Change-Id: Ida75392eefdbde900a620a3c6be0bb5ad31d6245