[roll] Roll fuchsia [kernel][phys] Truncate NVRAM with kernel.memory-limit-mb
Any portion of the NVRAM region that falls beyond the
kernel.memory-limit-mb cut-off should be discarded and become
inaccessible to the kernel. With http://fxrev.dev/1120681 we
forgot to update the logic around this truncation. Previously,
we would destructively discard available RAM, so that we could
just consult the last known RAM address at hand-off prep time;
since replacing that logic with an allocation of kTruncatedRam
type, we now need to consult the last known RAM address
specifically not of that type.

Original-Fixed: 369580523
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125783
Original-Revision: 6bda5b5176a507dfc146018105528d5529501358
GitOrigin-RevId: de4069c17474f496f09d147166efcc13bd76f520
Change-Id: Iaa4c3f01c2b80143ceb8fb67a53de3f35924e503