[roll] Roll fuchsia [assembly] Move product.base_packages to bootfs for no_image
When the product indicates that a filesystem image (fvm or fxfs) should
not be generated:
  platform.storage.filesystem.image_mode = "no_image"

Then all base packages should be moved to bootfs. In this situation we
assert that the product did not supply any config data for the package.

Original-Bug: 366062990
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118497
Original-Revision: 6d997c90f8f219cfeee560283f0be7100990d5b2
GitOrigin-RevId: c6fd4e6d5e75eaf1a45ad3bfbd445282205e8e4b
Change-Id: Ic288f277a05a906770290aa32f7da53cf1fa26dc