[roll] Roll fuchsia [netstack3] No route lookup during route discovery
When installing routes during IPv6 route discovery, skip checking if the
route was already present. Now that the routes are installed in a
dedicated NDP route set, this check is unnecessary.

The check was also inherently racy: Adding routes occurs in bindings,
so it's possible that bindings has a queued request to add the same
route that has not yet been actuated to core.

Original-Bug: 42079625
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123979
Original-Revision: 089110881380fe3a29f3da6e146398dcb54f0772
GitOrigin-RevId: fcc6d5116542416a1c533e68d3e1cabf22a40bfd
Change-Id: Id5c6c8e43d169d3359aa40f370253038c0ad5e76