[roll] Roll fuchsia [dns/netemul] Split DHCP test into DHCPv4 and NDP
Separate the existing DHCPv4 and NDP test into two separate tests. We
expect the Netstack3 out-of-stack DHCPv4 client to currently update the
list of servers used by the DnsResolver, but we don't expect to see NDP
messages. The existing test was failing but it was unclear whether due
to DHCPv4 servers not existing, NDP servers not existing, or neither
existing.

Create a library for the netcfg integration tests to expose a
`with_netcfg_owned_device` that can be used by other netemul tests to
simulate an environment where a device is discovered in devfs, picked up
by Netcfg, and then observed through the Interfaces watcher. This is
required for the DNS tests involving netcfg to return the correct
results.

Original-Bug: 328647117

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1125237
Original-Revision: ac9463d3e973876c08695859a180f3183ca4a6ec
GitOrigin-RevId: 5e6002d01fd70cd267200b0540133ae3b8c4954e
Change-Id: Ia8d2fb61d992c68a811623879efcd88ffd042d8f