[roll] Roll fuchsia [dl] Rename TraverseDeps to ReifyModuleTree
TraverseDeps is renamed into ReifyModuleTree, which will check if the
module's module_tree() is set before calling the enqueuing logic to
construct it.

The LinkingSession will use the root_module's module_tree() for the
subset of local-scope modules needed for relocation.

The RuntimeDynamicLinker will call ReifyModuleTree before returning a
pointer to module, if the module was already loaded.

Tests are enabled that require a loaded dependency to have a constructed
module tree when it is dlopen-ed.

Original-Bug: 354786114
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124681
Original-Revision: 3664747a5d33aa813da1c53336e04e2bb3b21434
GitOrigin-RevId: 8be7c310abe5e6b326ca73f6e082fd0e5d7d277d
Change-Id: Ib87cb0ee0807554dbc9b62f023628a54d9a94acb