[roll] Roll fuchsia [netstack3] Propagate controller removal to Core
Currently, when a fuchsia.net.filter/NamespaceController channel is
closed, the filtering resources owned by the controller are removed in
Bindings, and events indicating the removal are dispatched to watching
clients, but the removal is not actually committed to Core. The state in
Bindings and Core remains out of sync until the next commit to the
filtering state, at which point Bindings's updated state is committed to
Core in its entirety, resolving the discrepancy.

Instead, resources should be removed from Core immediately when a
controller is dropped. Flesh out the existing test that exercises the
controller lifetime semantics to actually verify the side effects of
filter installation and removal, by attempting to connect to a TCP
socket listening on a local port, both before and after the controller
is dropped (when the controller is alive, the connection attempt is
blocked, and when it's dropped, the filter is removed and the connection
attempt succeeds).

Original-Fixed: 369836595
Test: netstack-filter-fidl-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126720
Original-Revision: 4fc7b6960034d89e9c1e2d8709fd825a63e4fb95
GitOrigin-RevId: c72f8652ba3c99722184993c5d3d65dd3fdb13f7
Change-Id: Id84596ff25f6d49398ac75e3e19f19814a96ccfb