[roll] Roll fuchsia [owners] Remove kevinwells@google.com from src/sys/pkg/lib/sse/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/kevinwells@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127546
Original-Revision: de40d827248860808f74e024116b7b955e47003a
GitOrigin-RevId: 9eff0e351594c6a530d1d263830704db1704ffc7
Change-Id: Idcb4b05e519f5547709ebc2afdb97e834a122a2d