[roll] Roll fuchsia [kernel][pi] Remove lambda callbacks during PI propagation.
Refactor code in scheduler_pi.cc to remove the use of lambdas as the
main way to inject code into HandlePiInteractionCommon to deal with
the operation specific handling of updating a target nodes dynamic
scheduler parameters.  Instead, use a helper class along with a CRTP
pattern to achieve a similar result.

The CRTP pattern makes it possible to preserve the static thread
safety annotations on the operation-specific callback, making the
common handler aware of the callback's annotations.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093080
Original-Revision: 0cfd36bec7762263cb36198373792da571348459
GitOrigin-RevId: 3148be465a75e40a50b941601fc697665dade562
Change-Id: I996614354c6f543a5132eda61fbc2b4a4cc928f4