[roll] Roll fuchsia [intl] Always skip notifying nil changes
The intl properties provider could emit a change that is equal to
its actual default value. This is not a bug, but the behavior
is undesirable. So, prohibit it.

This does not affect usual operation but does make testing easier.

Original-Bug: 370076752
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128219
Original-Revision: a607ecd7fb920007b35e90c56dd3abe3669d4da0
GitOrigin-RevId: d689b94cc30c6e9a9a6041ca0d6606baf0d6239c
Change-Id: I5ba4daa164fabe0094ff24f99a3c6c3f9e237bca