[roll] Roll fuchsia [starnix] Use Starnix own address selection mechanism on bad mmap hint
If an address was passed as a hint, check if it is available, otherwise
ignore it. Previously, we were eagerly trying to map the hinted address
and retrying without it on failure, causing Zircon to choose the new
address.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127373
Original-Revision: 3da7461e63f3a0c795f7d67b597356ff2535f01f
GitOrigin-RevId: c3386d0eca62c8213162aa930467113b9bf3f621
Change-Id: I8c0488ab09539cdf79ba1ac0fa13e8fe04596b88