[roll] Roll fuchsia [netcfg] Remove input interface from masquerade API
Currently the fuchsia.net.masquerade/ControlConfig struct includes an
input interface, documented as "the interface carrying the network to be
masqueraded". However, in practice, this interface does not have an
effect on the implementation: the masquerade NAT functionality provided
by the netstack only allows matching on an egress interface, because the
ingress interface is not available in the POSTROUTING hook, where
masquerade NAT is performed.

The input interface was used to call
fuchsia.net.filter.deprecated/Filter.EnableInterface, but enabling
filtering on the input interface is neither necessary nor sufficient for
masquerading to work, as described above. Currently NetCfg *also*
enables the output interface, and that *is* load-bearing for
masquerading to be correctly performed.

To clarify the semantics of the API, remove the input interface field
from the fuchsia.net.masquerade API entirely and only call
Filter.EnableInterface on the output interface.

Test: policy-management-integration-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126674
Original-Revision: ee8ef7343719c4f58fca91eac2775d11c2f66be7
GitOrigin-RevId: 57b5bcb3891d0d156e76f0f9b935369e3191ae8d
Change-Id: I0b4f2afe5d6e008b89d7a9bfc5e0ab2ecb67d80e