[roll] Roll fuchsia [wlan][policy] removed unused past connection data
The `connect_attempt_time` and `time_to_connect` values that are saved
as part of `PastConnectionData` are currently unused. This change
removes them, to simplify some of the multi-bss support in state
machine. They can be added back in the future if there is a use for
them.

Original-Bug: None
Test: fx test wlancfg-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1123883
Original-Revision: 4c1c4a37a5a4333691ea17e619b0b150fb71f82f
GitOrigin-RevId: ee1837216322dfb249c712817fa3bd32e71efe87
Change-Id: I2d9d6edcad176bc1d20b08d1997a1a469f65c1e7