[roll] Roll fuchsia [owners] Remove glamrock@google.com from src/diagnostics/config/persistence/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/glamrock@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118542
Original-Revision: 46da577e704908f8adc43e684fc737e4f703b3ad
GitOrigin-RevId: f8e803247c792fda92f2dbdbf0534de0a6ce8cee
Change-Id: I3c7277684a78eca650e42d371b8c2f33a9537ce9