[roll] Roll fuchsia [pkg-cache] Clear variant when loading cache packages
We already do this for base packages, but not cache packages.
The full package url of cache packages are only used for inspect and
to serve the cache package index, where it clears the variant on the
spot, so it needed to clone the map.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126056
Original-Revision: 0864b5465639994a3a5de8c445dc37b6555087c5
GitOrigin-RevId: d3ada767d04bb84f0b05f517a7021ce38c557082
Change-Id: I828de4fe5841c1260385296e8cfebf946668929c