[roll] Roll fuchsia [zx][rs] fuchsia-zircon -> zx
This shorter name matches the common convention of
`use fuchsia_zircon as zx;` that is used so commonly in the
tree. Also aligns with the name of the C++ syscall wrapper
library libzx.

Also moves the library to a more suitable location for
eventual inclusion in a Rust SDK.

Leaves forwarding targets with the old names so that
the tree can be migrated incrementally.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128283
Owners-Override: Adam Barth <abarth@google.com>
Original-Revision: 9361643dceeaede2674552c6d6def07fddabc49d
GitOrigin-RevId: 86e50312a550f3d5840ef7ab060bdc8c04043294
Change-Id: I415f012fc0891a5c81dc08ed39f7e958ec65ffd0