[roll] Roll fuchsia [owners] Remove glamrock@google.com from src/security/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/glamrock@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118543
Original-Revision: 4922fb4fc9e7af3e4e63fbaa6e8e1253947bad37
GitOrigin-RevId: ca8252be4a447f119e5f536b30a1d65a6a49577e
Change-Id: I208cc35dacddf6261c6971a223358e178e9129d3