[roll] Roll fuchsia [starnix] Introduce new Unix Domain Socket
This implementation is based on the protocol defined in binder.fidl and
allows to use a unix domain socket to communicate between fuchsia and a
linux program in starnix and pass file descriptor through this socket.

This is an immodified reland of https://fuchsia-review.googlesource.com/c/fuchsia/+/1126412
The revert was only done to allow reverting a previous CL. This CL was
nto problematic.

Original-Fixed: 367349275
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129552
Original-Revision: 81bf67c7d209f183900268d1644810c436441aea
GitOrigin-RevId: a1dd7268c66d19e577ba59767e4da394b842d1d9
Change-Id: I57683e9b51619d45a2d08cadbe9340cbc39db693