[roll] Roll fuchsia [sdmmc] Clean up GN targets
Create a "source_set" GN target containing the source files and
dependencies of the sdmmc driver. Make the GN targets of the sdmmc
driver binary and sdmmc tests depend on this new target.

This makes it easier to ensure that the sdmmc driver binary and tests
have the correct dependencies.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1101035
Original-Revision: 01a38d32898c3b37f822540a370825287e44ce88
GitOrigin-RevId: 49c1a86b763a5177264dec089e60c6f1be738211
Change-Id: I607d07aa4be9aa9af02bf1b11ed8a020498967d1