[roll] Roll fuchsia [netlink] Decouple unit tests from full event loop
Decouples the netlink EventLoop into EventLoopInputs (which must be
provided up-front) and EventLoopState (which is computed from the inputs
after some async initialization and then is kept updated during the
whole event loop).

Introduces an `EventLoopComponent` enum that is similar to a `Result`
whose `Err` variant is only inhabited when `cfg(test)`. This can be used
to represent EventLoopInputs and EventLoopState fields that are always
present in production but may be absent in unit tests.

This allows us to run unit tests with only a subset of the full event
loop while maintaining type safety for the full event loop in prod,
thus simplifying the unit test setup and eliminating the need for
many of the complicated test features to serve no-op versions of
various previously-required FIDL protocols.

Multiply: netlink-lib-test
Original-Bug: 358649849
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127433
Original-Revision: cbe92e6092748881fbf629143962da827bc4860a
GitOrigin-RevId: a2e865457cb763e88b0092a04b04587d995347f6
Change-Id: I7febe9b353d6c48294c077fc9f412ab8358141f2