[roll] Roll fuchsia [testrunner] Allow /usr/share/ca-certificates in sandbox.
This is the directory where Debian-based systems store the
contents of the ca-certificates package which has the majority
of recent roots of trust for SSL connections.

For example, this is necessary to allow our symbolizer to
connect to our symbol servers during tests that want to
symbolize backtraces.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129752
Original-Revision: 6c3352bdc82961d6fc9edabe680f952c736c39f6
GitOrigin-RevId: 934b84838f63aa06f44412586156a7584bb470aa
Change-Id: Iab4097a5983afa999217e8e0a4b52cb39cdec9ac