[roll] Roll fuchsia [dl] Improve test logic for BasicGlobalDynamicTls
The previous version of the string matching only matched the glibc
format. The only reason this worked is because these tests were only
running on the host, and were not yet using the musl libc, since it had
a bug in its handling of undefined weak symbols.

Additionally, the host toolchains are able to select the TLS dialect for
either traditional TLS or TLSDESC. Because of those factors, we only
ever saw the glibc version of this error, until we tried to enable these
tests for Fuchsia targets in https://fxrev.dev/1111133, and had bots
running in CI that were expected to pass. What still isn't clear is why
some arm64 bots were able to pass this test (or why they skipped it?).

This patch updates the matching for the error string when
get_tls_dep_data is not available. Additionally, it incorporates related
test clean ups to the logic of the HAVE_TLSDESC block, which was always
silently skipping that code. Finally, it groups some of the checks more
logically.

Original-Bug: 338295240, 325518247, 342480690
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1118853
Original-Revision: ee924fee61e29404ed1ca822d16ad90eeba53d27
GitOrigin-RevId: 3b30aa35f58249ecb523884a37e4ef934ce31181
Change-Id: I75a17c335457f4c84525140af8ae4bdc2b9874f7