[roll] Roll fuchsia [kernel][timer] Add support for boot timers
Adds support for usermode timers on the boot timeline. Note that boot
timers do not currently wake the system if they fire during a period of
system suspension.

Also adds the clock_id to the zx_info_timer_t structure so that
userspace can figure out which timeline a timer is using at runtime.

Please read
https://fuchsia.dev/fuchsia-src/contribute/governance/rfcs/0260_kernel_boot_time_support
for more information about this change.

This change also adds a test for the behavior of boot timers, which
asserts that the timer fires during periods of suspension. This test
relies on zx_system_suspend_enter, which in turn relies on the CPU
system resource, so we needed to move the timer core tests into the
standalone-only mode. Doing this also allowed me to tighten up some
slack checks, as the standalone core tests do not set timer slack
policies.

Original-Bug: 328306129
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103275
Original-Revision: 7368a75b5f15101013decfecdd3166ae09884a89
GitOrigin-RevId: d22572b6b3235acbf40dbe7f74b54a0c4edc227f
Change-Id: I31e7111d73056a6e9047d9d02e1b71668b3f5174