[roll] Roll fuchsia [netstack3] Perform correct type of NAT in all hooks
Currently, only Redirect NAT (a special case of DNAT) is supported by
Netstack3. DNAT is only configured in the LOCAL_EGRESS and INGRESS
hooks, which is important because rewriting the destination of a packet
affects routing, so it must be done before a final routing decision is
made.

However, for reply packets on a connection that is DNATed, their source
address/port is rewritten to match the original destination of the
connection. This source rewriting must occur in LOCAL_INGRESS or EGRESS,
i.e. the same hooks where SNAT can be configured. Today, NAT rewriting
of packets in both the original and reply directions of a connection is
performed in the DNAT-specific hooks, which is incorrect (and can cause
more serious issues for, e.g., Masquerade NAT).

Change the filtering logic such that NAT is *performed* in every hook,
but can still only be *configured* for new connections in the
DNAT-specific hooks (LOCAL_EGRESS and INGRESS).

Original-Bug: 341771631
Test: netstack3-filter-test
      netstack-filter-fidl-tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129760
Original-Revision: 6a23cb4ee122cd9d3eb66e2b8990e1d4d6261b2a
GitOrigin-RevId: 8da93434b028b08dc168f10d9294bdeb24dda65c
Change-Id: I4cbb16374d3f0588f7cc8fc5e4731377cd6af1ea