[roll] Roll fuchsia [sdk][devices] Adding count field to cpu definition
This adds a "count" field to the cpu definition in the
hardware schema. This value is used by the emulator to
configure the number of CPUs to define when running the
Fuchsia instance.

Previously, it was hardcoded to 4, so as a backwards compatibility
method, the cpu.count field defaults to 4 if it is not present when
deserializing the data.

Original-Bug: 369114507
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1124052
Original-Revision: f7982ef9e904293d75f91426728db291010ad9a2
GitOrigin-RevId: ff9af8aae92a690431797a85a2da5dc0f2ad9cdc
Change-Id: Id2c2b2a5b5bca7bd8d6d1a9430b67ec6bbdb44e8