[roll] Roll fuchsia [kernel][crashlog] Fix reporting a disabled debuglog.
When a crashlog is generated, and the kernel requests that the
debuglog be added to the crashlog, whether or not that actually
happens is gated on the `render_dlog_to_crashlog` boot option.  When
the option was not set, the crashlog was _supposed_ to contain a clear
indication that the log was requested, but deliberately not rendered
because of the boot option.

This was not working correctly; fix that issue.

Original-Bug: 364882116
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130772
Original-Revision: 248b403841dfd83ba31e772262e6553cfba22ac7
GitOrigin-RevId: 848bb30a67a480fbaeef0a693adff471bdba00cb
Change-Id: I4d8791a3bbdd753977f3e6873aba5dbe46efee01