[roll] Roll fuchsia [starnix] Only remove the timer when deadline hasn't changed
Previously we would always remove the "same timer" from the heap, even
if the deadline for it had changed. This introduced a race condition
where:

  1. An interval timer was scheduled with deadline A.
  2. The "local" timer triggered, the outstanding hrtimer call was
     cancelled, and the timer's deadline was updated.
  3. The hrtimer actually had time to trigger, so the callback was
     called. This callback removed the timer from the timer heap, since
     it relies on the interval timer logic to reschedule the timer
     appropriately.

However, the timer had already been re-added (with a new deadline) to
the heap, so removing it caused us to remove the next instance of the
timer.

Longer term we likely need to come up with a different structure for
managing the timer <-> hrtimer relationship. We could also introduce an
API that is easier to use for us to schedule Fuchsia wake alarms in a
hardware-agnostic way. That component could then be responsible for
managing the actual hrtimers.

Test: This will be tested by the sysfs suspend tests in a future change.

Original-Fixed: 371006392
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130458
Original-Revision: cedae7f75bb8cf3cd9de61bf8990f23c628ade16
GitOrigin-RevId: 80ca3edc6a3681e99a898c93f81f58a9aa93004d
Change-Id: I301bcac3304b203e5fbdc9d520e317be59177605