[roll] Roll fuchsia [ffx] `ffx target list` should report port
When the port is not the default target ssh port (22), ffx target list
should display the port. Among other things, this will allow
programs that wish to provide the IP:port as the target (which
is a requirement for `ffx-strict` invocations) to use discovery
to find the appropriate port.

Original-Bug: 42154593
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129908
Original-Revision: 9ff7c574ea476818d2e5da0ad345c8b1a1eea956
GitOrigin-RevId: 87460c8f9af8a68b4ddbb39a9b97ddcaeeebef61
Change-Id: Icbc525e48c4e5ba1f76ed10f108a42035f527f6a