[roll] Roll fuchsia [ffx][repo] Add validation for having a target device
This adds a check before starting the server to make sure there is a
target device identified before starting the server. This check is
bypassed if the `--no-device` option is given since in that case,
the device is not needed.

Original-Bug: 370813472

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129857
Original-Revision: e21e33ed7186281f5071aafd89c89da2ca8cb0cb
GitOrigin-RevId: 94cd8abfc03409c3124110acd6404c3d1d6f14ac
Change-Id: Ifbabb2dd033308732bdce478541471264d726339