[roll] Roll fuchsia [ffx][fho] Exposing host_ssh_address for direct connection
This adds a method to get the host_ssh_address from the overnet
connection when running directly, without the daemon. This is the
address of the host from the device's perspective and is used to
address resources on the host, such as package servers.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130020
Original-Revision: da5bc72841170b4ce854898d05f2c6de7536cf46
GitOrigin-RevId: 8e24bd3619291eb3b3b8d8039e3c343ff75149ec
Change-Id: I4fd20f38c1fafb871a2b40fc85c446852357fdeb