[roll] Roll fuchsia [owners] Remove kevinwells@google.com from src/lib/testing/fuchsia-hyper-test-support/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/kevinwells@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127728
Original-Revision: 5975ccfb2c89285c6280b27f0ca77d4a502e541f
GitOrigin-RevId: 92b54c88482c24f3cbde8bd393ab96bd2ce238ff
Change-Id: I59413f35cdd7e809ab86622c195480b7d0637b09