[roll] Roll fuchsia [remote-control] Add default identifier
When used in embeddable configurations, rcs does not have access to most
of the protocols it uses to determine identity. As a result we need an
alternative approach to determining identity. For now, we can use a
hardcoded identifier. This change in the future once we have better ways
to determine uniquely determine identity of the system.

Original-Bug: 335549697
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128279
Original-Revision: 2562597fbcc10853cb10d97c2a4b4fab8771a3b8
GitOrigin-RevId: d30e16e3143c4806474beff5f5e9c079ed056ae9
Change-Id: Ib9db2f422c107424df33d74298a53e5eac407efc