[roll] Roll fuchsia [functionfs][adb] Add Proxy wake support
Add support for proxying the ADB channel so that new
information will wake the starnix container.

The actual proxying is disabled in this CL and will
be enabled later.

This also fixes an existing issue in the code where
events from the ADB stack were never read. They need
to be read once wake proxying is enabled. The code
spawns a thread which reads and discards these messages
in a loop.

Original-Bug: b/369114941
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1129772
Original-Revision: eb9e6b09067cbc336bd343201ad2b1279d16a498
GitOrigin-RevId: 4fedf466c9504f488cd4a20c170287d892ed0319
Change-Id: Ic82f1194b655d34bbe8f450e2d42d2b54adc6e53