[roll] Roll fuchsia [tee] Stub TEE_UnmaskCancellation
Instead of panicking, return true from TEE_UnmaskCancellation. This is
needed to pass preconditions for some tests, even though the tests
themselves do not rely on this functionality. Proper implementation of
Cancellation APIs is left as a TODO.

As a result of this change, regression1038 no longer panics and fails.
It talks to the TA successfully and determines to skip the test because
it doesn't support MTE.

Original-Bug: b/361406102
Test: fx test //src/tee
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132523
Original-Revision: 7e5bfb887caa3b88e71abc8b996d6a6238a78c32
GitOrigin-RevId: b8e6aa68cd080fee25f1bd81f4526d837d04fda3
Change-Id: I5232b824907c1753d1fbbe5549ff5fda8ca14ac3