[roll] Roll fuchsia [audio][drivers] Power down until first client
RingBuffer::SetActiveChannels is optional; audio drivers
must work well for clients who _never_ call it.

That said, for hardware that can be powered-down (and audio
drivers that support RingBuffer::SetActiveChannels for
clients to control the hardware's power), we would ideally
put the hardware into a powered-down state by default, until
the first client connects.

One that occurs, however, we want client changes to the
hardware's power level to be "sticky" and persist even as
RingBuffers are destroyed and others created.

With this CL, during initial configuration (bootup), we
power-down the hardware. Upon the creation of the first
RingBuffer, we automatically power-up. Subsequently, the
only thing that changes the hardware's power level is the
SetActiveChannels() call.

Original-Bug: b/371081956
Original-Bug: b/371088613

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130235
Original-Revision: 02f3c4243502d55c0e48033c80df6f8cfb9cce59
GitOrigin-RevId: 2ff880934ed8f5852cdf7bd958fac86884d1b944
Change-Id: I67732f877dc8a6e13a41ab7ff58107cc478bf2e9