[roll] Roll fuchsia [sestarnix] Remove FileSystem root node label workaround for "fuse"
Allow the root node of "fuse" filesystems to be labeled during
resolution of the filesystem's labeling details. Previously this
could result in a deadlock with userspace, if the filesystem was
marked for "fs_use_xattr" labeling, when attempting to fetch the
node's security attribute.  Now that "genfscon" statements in the
policy are respected, the workaround can be removed.

Original-Bug: 369328835
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134252
Original-Revision: 93e44d427cfa14fa829da3a7c682efdd89002089
GitOrigin-RevId: 8cfb223a3c423e314b38d078a91a0ac73e744bc9
Change-Id: I1efdbee80eb1b3fe9b2e05c0871fc6669518d535