[roll] Roll fuchsia [owners] Remove dworsham@google.com from tools/testgen/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1133331
Original-Revision: ee4157e0a4e7ad96520245a1ef87f49bb3033ee8
GitOrigin-RevId: 1e01326b91400d7dbe676e491a8b6a66da7df3bf
Change-Id: I45f4c67be6184a4067c132d0c40feba790428634