[roll] Roll fuchsia [netstack3] Move Mtu to SendOptions
Take the mtu argument from the IpSocketHandler API and move it to the
SendOptions trait. The option is only used in a single place: to limit
the MTU of sent ICMPv6 errors.

Also change the representation in SendIpPacketMeta to use the Mtu type
directly, and default to using a maximum u32 value everywhere. The IP
layer should always be enforcing the device MTU on egress, meaning when
we start doing so we'll apply a min operation between the device and the
requested MTUs.

Original-Bug: 371593049
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1130773
Original-Revision: 0d0b4f5733142f8c2415a38d9cf408538e444f96
GitOrigin-RevId: c57af8e2534864dc6f27e411e24def1ccf0b49cf
Change-Id: Idd11ca009beb67182fd8dcdfd7e7342b510a79a8