[roll] Roll fuchsia [owners] Remove dworsham@google.com from build/dart/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134593
Original-Revision: 515f973f1df0a2fff0f010f765956934041669cd
GitOrigin-RevId: 3e4206d281047c115f1c61843eb0b851e717f3dd
Change-Id: Ifc20795c207908c53421e11980ae3b250ec2899f