[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/fidl/fuchsia.ui.test.input/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134596
Original-Revision: 14d0b5559251d665e559e3ffa218b6a7a345cb8b
GitOrigin-RevId: 50f32cb2d98c5f1262773f8e07520177ed5530eb
Change-Id: I0de7a34aeb5d0a71908aaaf41809bdfd2794706d