[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/fidl/fuchsia.ui.policy/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134672
Original-Revision: f1ded520786300d9553f4b296a24402c132767fa
GitOrigin-RevId: 6dc25de36e012b0ec6b7fedbdb3bab09c42d66e8
Change-Id: Ib12114286ffb56e753e3d9c5f893c60e21d591ff