[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/lib/sys/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134578
Original-Revision: 6ba5e7a8ae7c003622e93b9b4b3b004fb59b141d
GitOrigin-RevId: f23195cbef57eaed57fc6f80b09b199367717c04
Change-Id: Ia51e3277dc1f7fa472ed14462eaa5da2c386a73c