[roll] Roll fuchsia [coverage][arm64] Test group for arm64 coverage builders
This is a trimmed down group that only includes tests that contribute to
extra source coverage when run on arm64.

We expect using this group to help improve performance and stability of
arm64 coverage builders.

This change only introduces the group, a followup change in infra will
switch arm64 coverage builders to actually use this group.

See linked bug on how I chose the tests that belong to this group.

Original-Bug: 369940767
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1126636
Original-Revision: 3969b2cbaefb48ff1a6654926d8b55229dcf9218
GitOrigin-RevId: 1baf5bf42cf45335e39b470497a408153ad8af83
Change-Id: I6e083db0220f29b1402b191a33444d6ed49dfe25