[roll] Roll fuchsia [tee] Introduce TA runtime hooks for resource management
This change introduces hooks to be called as prologues and epilogues to
TA_CreateEntryPoint() and TA_DestroyEntryPoint() for each TA,
respectively. These are intended to be part of a private contract with
the runtime to set up and tear down runtime-specific resources for the
TA (e.g., around protocol connection).

These hooks are intended to panic if something goes awry.

Original-Bug: 371213067
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135497
Original-Revision: 32b45eb9428bd1affd0ab96faadf7d7fe364ea64
GitOrigin-RevId: 1555057cf7fc6d80208b621f340605c3a2c9b227
Change-Id: I0c9573a6fcb0dced0176b2e34abe3fc634e40479