[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/fidl/fuchsia.ui.observation.scope/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134582
Original-Revision: 96da99cc2129b812bf516859e47635c2c8cbad36
GitOrigin-RevId: 9016263f1b4bff38b26f0834d80dd1d3e83d2bfc
Change-Id: I30f3de58f14dc96f2062015c4a23f44b4f5eeb34