[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/fidl/fuchsia.ui.app/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134577
Original-Revision: 59f9812d787ab2c0dde39259a95220663fa413d6
GitOrigin-RevId: a277c37325cda2cecd98c6cbca44dfd6e7aa4157
Change-Id: I9e51b12a81ee3c5f4d054a1d24d3b4cf04b147d1