[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/fidl/fuchsia.ui.views/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134595
Original-Revision: c37c85ebcad082f10f8dd90ab54926b0b613eff6
GitOrigin-RevId: 911d280e5349789042dc1659f5871e073471cd60
Change-Id: I3fa3430ae9dccce43effedd931af2b5202037dd8