[roll] Roll fuchsia [components][sandbox] Replace program_output_router with direct lookups
program_output_router is program_output_dict wrapped by a Router. It is
an argument to build_component_sandbox used to fulfill routing requests
from "self".

This layer of indirection is not necessary, Instead, we can lookup the
Router in the program_output_dict directly, and use that.

Also: Add a dictionary test to exercise more of this logic (hopefully
this can catch a reference cycle, if there is one)

Original-Fixed: 361653724

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1106192
Original-Revision: 24d5d9a4604830e2a1bc9133350c71ad17e119d9
GitOrigin-RevId: 0dfbfa01675bab0249580f601f98854a53243d88
Change-Id: Ib34e55f5e2d47eef86664c4b8fc9975c7d80e0f8