[roll] Roll fuchsia [cpu_profiler] Robustly Handle Child Jobs Exiting
The profiler was flaking some e2e tests:

In the logs we see:
```
17:21:32.089411 [00055.098853][profiler] ERROR: [src/performance/experimental/profiler/targets.cc(227)] failed to get job: 88854: -25 (ZX_ERR_NOT_FOUND)
17:21:32.089711 [00055.099276][profiler] ERROR: [src/performance/experimental/profiler/targets.cc(233)] failed to make job_target: -25 (ZX_ERR_NOT_FOUND)
...
17:21:34.094972 [00057.103174][profiler] INFO: [sampler.cc(225)] Stopped! Collected 0 samples
```

What appears to be happening is that while we're traversing the root
job, in the time between us querying for a job's child jobs and then
querying the job itself, the job exits on us. Since it's a timing race,
that would explain why it flakes. Different usage patterns and whether
there is a short lived process/job running at the same time as the test
(which is e2e, and not hermetic) would explain why it could suddenly
start flaking a lot.

This patch includes some changes to MakeJobTarget to skip failing
children, reducing it to a warning from an error, and only fail making a
job if the job itself disappears from under it.

This also adds a test that spawns and destroys a bunch of jobs and
ensures we still successfully configure.

Original-Bug: 371463369
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135503
Original-Revision: 489ee380ef3c6b24a6e9b3453cc9989643b480cf
GitOrigin-RevId: 949cfc543b3b12ab1925c0906b8d679c041cf8be
Change-Id: Ibf5654a18cadcd57d9c2cd3c70764f3f714fb9ab