[roll] Roll fuchsia [debugger] Add DebuggedJob
Add a new DebuggedJob abstraction analogous to DebuggedProcess. This
class will be responsible for monitoring for job exceptions and serving
as an intermediary between the platform specific JobHandle
implementations and DebugAgent.

An additional method is added to JobExceptionObserver which will be used
when we have support for attaching to the non-debugger job exception
channel.

Also adds a new flag to ExceptionNotification to let the front end know
that this is an exception from a Job.

This is not hooked up yet, so no behavior changes are expected.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134738
Original-Revision: 2629c35207356c7e4a1b83c7e5b609857d10e20d
GitOrigin-RevId: 01d887a8a8ec6dff83915a1a0790d094f25ba772
Change-Id: Ica4dc2da545dd777f86827a1f9ea576c1bd076fd