[roll] Roll fuchsia [owners] Remove jbauman@google.com from src/lib/vulkan/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jbauman@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135431
Original-Revision: 6ed135b57624fd6e8fc9a1ff15ae5ae8cd1e499a
GitOrigin-RevId: f6d153eff743ab72db3b6dae437adb25682bdf54
Change-Id: I0093b4753e7a7d513721ccede538dc18b27f8805