[roll] Roll fuchsia [starnix][aio] Fix AIOTest.CloneVm
This test was flaky because we were trying to hold onto the memory for
the asynchronous IO operations with a WeakRef to a Task. In this test,
the task that submits the operation can exit before the operation
occurs. Instead, we need to hold a strong reference to the MemoryManager
for the task, which can persist throughout the operation.

Test: AIOTest.CloneVm
Original-Fixed: b/372774866
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136332
Original-Revision: f2a7ec8f57b714c39fa0734d56c7c61d65d4c1bf
GitOrigin-RevId: 3491cd2926ef0db59af2967eee0a7cccd5935cda
Change-Id: If6f457bacb7407c7662215b3247e9ed0a58fe9bd