[roll] Roll fuchsia [starnix][functionfs] Handle R/W on separate thread
Handle the ADB calls on a separate thread so we can use an
async context. The async context is important so we only go
to sleep after we have already queued a read, so we can be
sure we will be woken up when the read returns.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136173
Original-Revision: 481cb4a2d548c104ff21c19503bc1366e8ddba10
GitOrigin-RevId: de997ec3086b4993217afae3400f2cff2fa6c9a1
Change-Id: Iff5b8c7eacae3bcf067e580c60727186572b0e53