[roll] Roll fuchsia [netstack3] Emit the default rule through watchers
Before this change, we never emit the default rule that points to the
main table. This is hard to hold when syncing rules to Core. Instead
of hiding it, we expose this rule at the lowest priority.

Also made a small change to net-cli so that the rules are sorted by
the priority and index.

Original-Bug: 357858471

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1132379
Original-Revision: 2d9e48de986f67f4e480fa6d341a3d3db036debc
GitOrigin-RevId: e8109873a06560be11e88bf6a28ead4481d3bfcc
Change-Id: I5db04ded8429d31680ee284d000b9152c6e667cb