[roll] Roll fuchsia [fshost] Move route to correct location
Routes need to be included in the shard that includes the component
using the capability, not the one that declares the component which
provides it. Otherwise we are always forced to include client component
whenever we include theserving component.

This particular route broke microfuchsia as it doesn't include fshost,
but does include the driver framework.

Original-Fixed: 373387565
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137333
Original-Revision: 47bfcb9da6a9cd9e706efb7dc25639031318e76c
GitOrigin-RevId: 861413a59200152871175d239811613b579dffb9
Change-Id: I1670ef64e1d94e8f92eae63be4a34c26c7389177