[roll] Roll fuchsia [starnix][aio] Respect max_operations
Previously, we were counting the maximum number of operations for read
and write operations separately. This CL restructures the internals of
the AIO structures so that we apply the max to the sum of the number of
read and write operations.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137199
Original-Revision: dcab307a2af2a7bb94ebf16e1938507dc0ccf702
GitOrigin-RevId: d980debd5e7082c4c78de7c4e0fb0025c97b224a
Change-Id: Iff8b45f16deed7ac56f9d9afc4d27c1c73334347