[roll] Roll fuchsia [zx][rs] Type-safe exception reports.
Callers no longer need to use unsafe to interact with the union.

Removes the body of the Debug impl for zx_exception_header_arch_t
since that type is publicy constructible and we can't guarantee that
it is always valid.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135007
Original-Revision: 82ecd7e50200cbc87ec3714392a81de7fe91b04b
GitOrigin-RevId: d9d6cc148b7d8d4075d43ef1b85a05d11d5fc566
Change-Id: I0bf75f9812be31d3e1d7ac8b0e1a8c8836730380