[roll] Roll fuchsia [owners] Remove rosasco@google.com from src/lib/vulkan/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/rosasco@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136888
Original-Revision: eef1a8a721ff6b125a28fc9b921f7ea715f1a0ec
GitOrigin-RevId: a20530ae4d89599beded666bd5b46262e94009a5
Change-Id: Ia2259f293c650a86d3ce098680bd470c67297669