[roll] Roll fuchsia [owners] Remove dworsham@google.com from sdk/fidl/fuchsia.ui.annotation/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/dworsham@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134598
Original-Revision: 4122caf868eee1d587870d00ab1112d42d5f762e
GitOrigin-RevId: b5db9e914eb71073112aa772d91df1a4d9cd923f
Change-Id: I1ab605037d1242d42f54d084a2139699f138a29d