[roll] Roll fuchsia [owners] Remove jbauman@google.com from src/media/drivers/amlogic_decoder/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jbauman@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136394
Original-Revision: 796f7942b1957ed1b0748ca21fdc62f9d2b6d56a
GitOrigin-RevId: 6dafb8960e20076897f4d0e3f2504cf96cbd4354
Change-Id: I7f56b04c120cbb98303511e3b96b229af56b62a1