[roll] Roll fuchsia [owners] Remove jbauman@google.com from zircon/system/ulib/image-format/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jbauman@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136393
Original-Revision: 84600f3ba132384106301937103ad2afd89897a0
GitOrigin-RevId: 8bd0e6e6f0ca20f01d9243f9c59b8124158e587b
Change-Id: Ic243e46fd04e53fb8c659b0323eb38c69422ac3d