[roll] Roll fuchsia [owners] Remove jbauman@google.com from src/media/codec/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jbauman@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136392
Original-Revision: d9e463b98817daab7439805ca22691ba9a71512e
GitOrigin-RevId: 1f7d27542cc18a01d67984c42c86fef288c0cc57
Change-Id: I434fc3fb6d1f9fd83dcecc6e5f7ab296e5ca7a28