[roll] Roll fuchsia [owners] Remove jbauman@google.com from src/media/third_party/chromium_media/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jbauman@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135430
Original-Revision: ae5485c7bdd507b7cc41c57596cd71e232758c0b
GitOrigin-RevId: bfa65f3cf258f677218e920a896875dbb4652ff1
Change-Id: Ia3bee5193af397ca86ad90d479ba295c0bcb2eb2