[roll] Roll fuchsia [owners] Remove jbauman@google.com from zircon/system/ulib/sysmem-version/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/jbauman@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1134828
Original-Revision: dfe02d81e7841f1c8f37bb0bf271bbfcb7ce73e3
GitOrigin-RevId: 94c161694ea12a64784e9adf50858e874408d9d4
Change-Id: Ia9d7455f9f7fd97a2cd92ebaf1a845e0652bc45e