[roll] Roll fuchsia [starnix][ltp] Implement CLONE_PARENT
This CL implements the CLONE_PARENT flag for clone, which is tested by
clone08. We now pass an additional subtest of clone08, but we do not
pass the whole test yet because we fail a subtest of another feature.

Test: clone08
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137813
Original-Revision: 2e5b85a1267a0560ce45a5d37fd574335824a0dd
GitOrigin-RevId: d119dba4260a7249ec1d0e1ade0c2f04738ef810
Change-Id: I3431068b033cfadea87b1bcb516e0ad40ce421de