[roll] Roll fuchsia [bazel][sdk] Fix relative path product configs
If a fuchsia_product_configuration uses relative_path=True, the paths
for the base/cache/base_drivers were not relativized properly.

This CL fixes the issue and adds a test for the rule.

It was found that the rule sorts the product config for
relative_path=True, therefore the input and goldens were all sorted.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137521
Original-Revision: 1c347c86081cbe6d9c2877282ba972d9e7040ed2
GitOrigin-RevId: a9031e2ac10332ad8d23ed3a5e95bedf08cd7b58
Change-Id: I6b0c65798668a5ce817ec5afb9973c8085afc9df