[roll] Roll fuchsia [terminal] Move packages from legacy bundle to assembly product definition
This moves many of the packages listed directly in the gni file
to the targets in the build.gn file, nearly removing all of the
packages from the cache packages set.  The remaining packages
are in groups that will each be moved in a single cl for clarity.

Run-All-Tests: true

Original-Bug: b/361348304
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136090
Original-Revision: 7a7029ebfcf570997cb87b77eeba2ad991644f59
GitOrigin-RevId: 7b55ead8d9c04364417b832fbe3e74e219adac4d
Change-Id: Ia18d65bc3972a1e6fa9077415f81a7de4a155256