[roll] Roll fuchsia [windowed-stats] Implement event reactor combinators.
This change introduces the `event` module and an API for constructing
reactors that respond to data and system events. Reactors are
constructed with combinators that respond differently to events, such as
sampling user data with a `TimeMatrix`. See unit tests for examples.

To support this change, `TimedSample` has been generalized and renamed
to `Timed` (there are now timed samples and timed events). The
`DataSemantic` trait now has an associated `Metadata` type. Metadata
APIs are incomplete and this change assigns the unit type `()` to
`Metadata` for all `DataSemantic` types.

Original-Bug: 350797382
Test: unit tests
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1127883
Original-Revision: efdb86a2bb284eb1d47db3fba563ad16ed1bad93
GitOrigin-RevId: bde66762559050e14b08beb847663ece8b8aaac6
Change-Id: I12c813813b0bb681cf9dd7929f1b86665a1d86d6