[roll] Roll fuchsia [gpio][devicetree] Add controller index property for gpio-init nodes
gpio-init composite node specs already have a bind rule for the
controller ID to ensure that they bind to to correct GPIO controller.
This change adds GPIO_CONTROLLER as a bind property as well, using the
index of the GPIO controller as found in the pinctrl-0 property. This
prevents ambiguous bind rules in drivers that need to bind to multiple
gpio-init nodes.

Test: gpioimpl-visitor-test
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1138316
Original-Revision: 4edd999bfd81295c4da0bc984d06ee18fc71f8ce
GitOrigin-RevId: a2efbcafdc70a3027861792ff22bb875a3038fd1
Change-Id: I4bad456c718094a032f8841a7cdf27f1581aaa3d