[roll] Roll fuchsia [tee][runtime] Better handle empty VMO parameters
The runtime parameter translation logic from FIDL would unconditionally
map the VMO associated with a memref parameter even when the VMO was
empty, which is illegal.

However, we shouldn't just provide a buffer address of nullptr in such a
case. A zero-length buffer at a valid address is totally fine and that
should be presented to the TA as valid memory. Accordingly, we handle
this case by mapping in a zero-filled page from a fresh VMO - not
trusting that the original VMO had non-zero content - and specify a
buffer of size 0 rooted at that address.

Original-Bug: 371213067
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137952
Original-Revision: 004c56cad770743ee76ea0f86bf6d7f301bef114
GitOrigin-RevId: 44671f677e8b231d9abd73d80f5257da2f700031
Change-Id: Idb00629dfa2589687de0375acb3e66101049621d