[roll] Roll fuchsia [owners] Remove rosasco@google.com from src/developer/ffx/plugins/agis/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote the lowest CR value
(i.e. reject the CL) and the bot will abandon it. Vote the highest CR to
approve this change. You may also abandon this change.

See the owner's recent activity for context:
https://fuchsia-review.googlesource.com/q/rosasco@google.com

To report an issue, file a bug in the Infra>Codereview component.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1136887
Original-Revision: afd4174280990b5bcba387f1590239c745848ed4
GitOrigin-RevId: 2362a1fc1e7a47965dd4a72bb96c8144bbfa6b48
Change-Id: Ia5dafe22f6e16b43df8eba42559e97da281b68bc