[roll] Roll fuchsia [driver][compat] Remove TopologicalPath in compat
Removes the GetTopologicalPath method in the legacy driver compat
protocol. Makes the node passthrough implementation the one that is
always called instead.

This keeps the protocol in the fidl as we need to soft migrate the
vendor repo.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1135912
Original-Revision: 5b31c91b40e4c4cf41718bfe52b2ae677f20ee04
GitOrigin-RevId: 72d7914e889c311602dde5ac9b5fc75884f5cfcd
Change-Id: I1f179b402ba4baa5c6624f1824b38133860af141