[roll] Roll fuchsia [diagnostics][log] Improve severity API
Improves DX a bit instead of requiring the FIDL type to be part of the
public API. This improves how we can move to also support the log crate
as well as provides some improvements for the follow-up simplifying
pieces of the log library. There'll be more follow-ups to remove
redundancy.

Original-Bug: 344980783
Original-Bug: 364293789
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1103655
Original-Revision: 07e5deb9159b07c07001a2ddfec5cb2daba4bbea
GitOrigin-RevId: 5f49a6e5ec94a6d95adba731ee011c08927af3a7
Change-Id: Ifa90024e8edb2dcd1fddd783a66e8fd0dd3cc256