[roll] Roll fuchsia [cpu_profiler] Increase Resilience when Starting
Between configuring and starting, we may lose one or more of our targets
because they exited. This patch decreases previous error when a process
or job was not found to a warning and allows us to carry on if we fail
to attach to a target.

Original-Bug: 371463369
Multiply: cpu_profiler.sh

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1138265
Original-Revision: 8e2c1c4d36ee9afef286f2e976ac1a002dd2afae
GitOrigin-RevId: 97c16b113fd0e158c2c323bcfeb19b0e48adb545
Change-Id: I7b4d0015c90c695a196e32a921d3adb0b753d06c