[roll] Roll fuchsia [debug-dash-launcher] Use the bootfs package resolver
The launcher now uses the recently exposed bootfs package resolver to
resolve fuchsia-boot:// package URLs.

The tools trampolines are stored in directories named after the full
URL (with / replaced with | and : replaced with ^) instead of just the
first segment of the path now that it is easier to have packages with
the same path but different scheme, host, etc. This is user visible
but clients should be using the tools binaries via $PATH which is
unchanged.

Original-Fixed: 329167311
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1137867
Original-Revision: 1ca8751cab823c78b69f408a26a03b027646d4cf
GitOrigin-RevId: 0bc5562cd33f0d5832b1e7ee01e89055ca1299bd
Change-Id: I27dd71702c79e690ff45efa6cd4803c471e4dc81