[roll] Roll fuchsia [sdk] Move //build/bazel/bazel:idk to //sdk:bazel_in_tree_idk
This renames and changes the location of the IDK that is built
to populate the content of the @fuchsia_sdk repository in
the platform build's internal Bazel workspace.

The new name is clearer about its intended usage, and the
next location ensures that all IDKs are defined in //sdk/BUILD.gn
which makes things more consistent.

There are still some extra validation targets under
//build/bazel/bazel_sdk, but I prefer to let them here since
they involve both GN and Bazel graphs.

There is a new //sdk:bazel_in_tree_idk_tests target that
runs the Bazel SDK test suite directly against this IDK.

It is not necessary to add this to regular tests, as we
already run the test suite on the @fuchsia_sdk repository
itself, which is what really matters for the build.

+ Rename `@fuchsia_idk` repository to `@fuchsia_in_tree_idk`.

Original-Bug: 367274537
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1140000
Original-Revision: 3b851029136cbe953534aabb7ee574a6236ce9ac
GitOrigin-RevId: 6c44235c50a6d629ac82db22ec53b6113d6e0489
Change-Id: Id04bd7bb45ee9972ea1364e12b8fcc3bbe5caf69