[roll] Roll fuchsia [docs][stop_idle] Update to new sandbox API
Update the docs to pull from a rust file so we can be sure that
the code compiles and it will be updated when the APIs change.

Update some of the BUILD files to use the right target names
to make building simpler.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1138292
Original-Revision: d7cbe6ab2eba968691daff34084aff363f8c2343
GitOrigin-RevId: afb3a8b175081f8b7bde9c6aa8d391afb2a1f298
Change-Id: I0eef10a24565b2e6b7957743f8dfef7d3aca3e54