[roll] Roll fuchsia [build][sdk] Internal doc for IDK Bazel migration.
Provide a documentation describing the various parts, related
to IDKs and SDKs, that the platform build needs to manage,
and how they are going to evolve.

This document will evolve as migration progresses but is
useful to prepare interested developers to the upcoming
changes.

Original-Bug: 367274537
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141372
Original-Revision: 259d3837421c3cd5d1a103893ed9dfce251c7ffd
GitOrigin-RevId: 0b2baca2618715ff9a44136ea97416aca7c95d63
Change-Id: Idd8336ed8cce3e15140625e7ef549bd489dfcc6b