[roll] Roll third_party/go [go] Only define runtime tests in the fuchsia toolchain
This is step 1 of 3 of a restricting of the Golang Runtime to
only fuchsia binaries, as the last remaining host binaries
are tests that are also run on fuchsia.

This step removes the usage of 'use_prebuilt_go', letting
the toolchain select it appropriately (true on fuchsia, false
on host).

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/third_party/go/+/1142621
Original-Revision: 94e10487ab9431e7a561b305a3b8bc59e590783f
GitOrigin-RevId: 91e09b3930d762b240473b291a2e4cb6e8306878
Change-Id: I07cc035f9b0760fc1eea7dc62ec34c820387cd2a