[roll] Roll fuchsia [ffx][config]: get function is no longer async
The ffx_config::get function was needlessly async, now make it sync.

Note: any functions that needed to be async only because they
called ffx_config::get have not been migrated to be sync; that is left
as an exercise for the authors.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143253
Original-Revision: 443dcf221c7ee64132b25923badc82f0a6882fbb
GitOrigin-RevId: e9763935c7b17970e74b1e71c075c14660c8cd70
Change-Id: Ia6b93a63c52c9865528704bcd0e870264d9780f1