[roll] Roll fuchsia [hrtimer][vim3] Fix Stop API IRQ handling
Remove calls to cancel IRQ handlers in the FIDL Stop() API
and starts them at initialization.

Canceling an IRQ handler can race with an incoming IRQ,
which could then become pending and never be acknowledged.
The reason the hrtimer canceled IRQ handlers was to
prevent an incoming IRQ from generating a notification after
the FIDL Stop() API was called.

Hence, we change the FIDL Stop() API removing the call to
Cancel() and we update the IRQ handler method such that
it acts only on started timers and valid IRQs.
If we get an IRQ handler call after we have stopped the timer
we'll log and ack this IRQ but not attempt to restart the timer
(normally the IRQ handler will not be called since Stop() stops
the timer hardware).

For the Cancel() on shutdown, added a warning on error and
a TODO for a generalized solution to lost IRQs in that case.

Also improved logging and removed unused code.

Original-Fixed: 369886005
Test: fx test -o aml-hrtimer-test
Multiply: aml-hrtimer-test

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141433
Original-Revision: 85544780b621ac4adf089bcbbc0a587334f646c6
GitOrigin-RevId: f01b996597ef314e48be1037e7950c3570d7f09a
Change-Id: I31f6cc706f83a51b5769bfbb446d5108a64bb0a2