[roll] Roll fuchsia [wlan] Refactor SME roam result handler.
This is a refactor of SME roam result handler logic into a
dedicated function. This refactor will allow Policy-initiated roam
attempts to use the same result handler used in Fullmac-initiated roam
attempts. There is no behavior change expected from this CL, as it is
a refactor to reduce cognitive load for the Policy-initiated
implementation CL to follow.

This test passes existing unit tests for SME, and the Antlion suites
BeaconLossTest, VapeInteropTest, WlanPhyComplianceABGTest,
WlanTargetSecurityTest, and WlanWirelessNetworkManagementTest (with a
device running BTM capable firmware, and also with a device with BTM
disabled).

Test: fx test wlan-sme
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1128755
Original-Revision: 3676763bd7d95e3748141ca18a98169d807e6e2f
GitOrigin-RevId: 58335dc0a72acab737d0fe53ad7d955f2e09beee
Change-Id: Idc54d73605d73ab4a008064744ca7156f1507e06