[roll] Roll fuchsia [go] The go runtime and fidl-enabled go is fuchsia-only
This is step 2 of 3 of a restricting of the Golang Runtime to only
fuchsia binaries, as the last remaining host binaries are tests that
are also run on fuchsia.

This restricts the use of the fuchsia go runtime to just fuchsia
binaries, and removes the last host-side tests that were using
the fuchsia Go runtime.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141944
Original-Revision: 4e926a61aba89cc540d04e3e74801242b51fe2d8
GitOrigin-RevId: b993bc8b49fe4732c5c31ab2a9f37614670e6c4b
Change-Id: Id8e941f777130d72e01e57a2791fd53b16856fc2