[roll] Roll fuchsia [arm][timers] Be consistent about counter selection, prefer VCT
Prior to this CL, there was a small disagreement between the kernel
and user-mode when it came to ARM system counter selection.  User-mode
_always_ used the virtual counter (VCT).  The kernel, however, would
choose between VCT and PCT (the physical counter) based on which timer
interrupts were offered to it by earlier stages of booting.  If a
physical timer interrupt was provided, PCT would be used, even if a
virtual timer interrupt was also supplied.

This CL changes two behaviors in order to keep everything consistent.

1) It changes the kernel to prefer the VCT over the PCT if given the
   option to do so.  This should ensure that when Zircon is booted as
   a guest OS, that the host hypervisor can use the EL2 offset
   register in order to control the guest's perception time instead of
   needing to trap and vmexit every time the counter is read.  This
   said, if no virtual IRQ is provided, PCT and the physical timer
   hardware will be used as a fallback.
2) It changes the VDSO given to user mode to use the PCT as the
   counter reference if that is what the kernel was forced to use.

Finally, this CL addresses an edge case at the start of boot. Some
Cortex-A73s have an error in them which can sometimes lead to a bad
counter read (either PCT or VCT), and requires a special double-read
protocol in order to work around the issue.  Before this CL, we would
start the system by assuming that we did _not_ need to implement the
protocol during counter reads.  As the CPUs each came online and their
cores were identified, if any A73s were detected, we would switch to
using the A73 workaround.

This CL reverse that behavior.  Now, we start by assuming that we will
need the A73 workaround, and switch to not using the workaround only
if we can demonstrate that none of our cores are A73s.  This means
that we will be paying the penalty of needing the workaround until all
of our CPUs have manged to identify themselves as not being A73s, but
we will not run the risk of a bad counter read by assuming that we
don't need the workaround, only to discover that we actually do later
on.

Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1131295
Original-Revision: 83e0b69327db7ad037d419ba5ced8fe26753982d
GitOrigin-RevId: f6906fb4f02e97e03a903971e3d9e0940a05eef9
Change-Id: Id3a8f8a0f0efd1aff00207591dfac692f1213fb5