[roll] Roll fuchsia [build][sdk] Detect SDK category violations for "partner_internal"
Previously, the `detect_violations()` Python function was not run on
FIDL libraries in `_sdk_partner_internal_fidl_libraries`. This allowed,
for example, FIDL libraries in the "internal" category to be included
in this list. See https://fxrev.dev/1139635 for an example of this.

Original-Bug: 366221510
Test: Revert https://fxrev.dev/1139635 and observe that `fx build` fails due to a category violation.
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1143480
Original-Revision: 742cba56607761b1ea1546c75e1951b77dc28fd7
GitOrigin-RevId: 16b7a1064a1957686dd071dbea5c0774fd52646a
Change-Id: Ifef38816d25aba72c88eba8f78abef6ae887945d