[roll] Roll fuchsia [ffx][usb_fastboot_discovery]: skip version check
Now opening an interface does not require the fastboot version variable to be 0.4 exactly.

As long as the getvar command returns either an Ok(Reply::Ok) or Ok(Reply::Fail) it is valid. It implies that the fastboot protocol is up and running and ready to communicate.

Original-Bug: b/355707967
Original-Fixed: b/355707967
Original-Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1141152
Original-Revision: e6198107cc059c615d7b3599573da7d260027b13
GitOrigin-RevId: 545750221f05ede1d86a9feea6b080d2c9229156
Change-Id: I91d4117dd2001fbc77a6778f29ca38214f5605c8